pyproject.toml
[project]
name = "poketerm"
version = "0.1.0"
description = "A CLI tool to get random pokemon ASCII art and pokedex info"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "pillow>=12.1.1",
    "requests>=2.32.5",
    "rich>=14.3.3",
]

[project.scripts]
poketerm = "main:main"
